app_lake_eo/requirements.txt
python-dotenv
databricks-sql-connector
databricks-sdk
fastapi
uvicorn
